+++ opendut-edgar/src/common/settings.rs
@@ -2,6 +2,8 @@ use std::path::PathBuf;
 
 use opendut_util::settings::LoadedConfig;
 
+pub const CONFIG_APPLICATION_PREFIX: &str = "edgar";
+
 #[allow(non_upper_case_globals)]
 pub mod key {
     pub mod peer {
@@ -25,5 +27,13 @@ pub fn default_config_file_path() -> PathBuf {
 pub fn load_with_overrides(overrides: config::Config) -> anyhow::Result<LoadedConfig> {
     let edgar_config_hide_secrets_override = opendut_util::settings::Config::default();
 
-    Ok(opendut_util::settings::load_config("edgar", include_str!("../../edgar.toml"), config::FileFormat::Toml, overrides, edgar_config_hide_secrets_override)?)
+    let loaded_config = opendut_util::settings::load_config(
+        CONFIG_APPLICATION_PREFIX,
+        include_str!("../../edgar.toml"),
+        config::FileFormat::Toml,
+        overrides,
+        edgar_config_hide_secrets_override
+    )?;
+
+    Ok(loaded_config)
 }